Use size_t for the number of shared memory bytes.

+++ hoomd/mpcd/SlitPoreGeometryFillerGPU.cu
@@ -170,7 +170,7 @@ cudaError_t slit_pore_draw_particles(Scalar4* d_pos,
 
     unsigned int run_block_size = min(block_size, max_block_size);
     dim3 grid(N_tot / run_block_size + 1);
-    const unsigned int shared_bytes = (unsigned int)(num_boxes * (sizeof(Scalar4) + sizeof(uint2)));
+    const size_t shared_bytes = num_boxes * (sizeof(Scalar4) + sizeof(uint2));
     kernel::slit_pore_draw_particles<<<grid, run_block_size, shared_bytes>>>(d_pos,
                                                                              d_vel,
                                                                              d_tag,
